+++ Valtuutus.sln
@@ -35,8 +35,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Valtuutus.Data.SqlServer.Te
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{62CDC085-2DF2-4389-B65B-CB893B15A2C1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Valtuutus.Data.Tests.Shared", "tests\Valtuutus.Data.Tests.Shared\Valtuutus.Data.Tests.Shared.csproj", "{9FA14311-F75A-473E-9042-E9A3F9187300}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "github", "github", "{BB859DDD-7E7F-4F97-B033-635432F127A5}"
 	ProjectSection(SolutionItems) = preProject
 		.github\workflows\build.yml = .github\workflows\build.yml
@@ -110,10 +108,6 @@ Global
 		{954D9B3C-A6BF-4211-9EB5-B205093593A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{954D9B3C-A6BF-4211-9EB5-B205093593A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{954D9B3C-A6BF-4211-9EB5-B205093593A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9FA14311-F75A-473E-9042-E9A3F9187300}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9FA14311-F75A-473E-9042-E9A3F9187300}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9FA14311-F75A-473E-9042-E9A3F9187300}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9FA14311-F75A-473E-9042-E9A3F9187300}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2612B20F-43DB-4486-AD80-1655A6A75586}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2612B20F-43DB-4486-AD80-1655A6A75586}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2612B20F-43DB-4486-AD80-1655A6A75586}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -141,7 +135,6 @@ Global
 		{C5EA2815-117F-4FF9-B6DA-87EC47B884C2} = {D383AD39-3BE1-43F8-BF40-E3F7A6D1CD1D}
 		{954D9B3C-A6BF-4211-9EB5-B205093593A6} = {D383AD39-3BE1-43F8-BF40-E3F7A6D1CD1D}
 		{C45DF205-1484-49D5-A8C0-5AEADCC144D3} = {62CDC085-2DF2-4389-B65B-CB893B15A2C1}
-		{9FA14311-F75A-473E-9042-E9A3F9187300} = {D383AD39-3BE1-43F8-BF40-E3F7A6D1CD1D}
 		{BB859DDD-7E7F-4F97-B033-635432F127A5} = {D7A6602B-1ADA-48CB-B6BC-DD7BDB273068}
 		{CE52FA1C-E7A6-4D10-A446-1D9209C79DAF} = {D7A6602B-1ADA-48CB-B6BC-DD7BDB273068}
 		{C796EA04-CF5F-4DCE-B37D-D59D611CC020} = {D383AD39-3BE1-43F8-BF40-E3F7A6D1CD1D}
