+++ tests/Valtuutus.Api.Tests/verify/ApiSpecs.ApproveCore.DotNet8_0.verified.txt
@@ -155,13 +155,13 @@ namespace Valtuutus.Core.Engines.Check
         public CheckRequest() { }
         [System.Diagnostics.CodeAnalysis.SetsRequiredMembers]
         public CheckRequest(string entityType, string entityId, string permission, string? subjectType = null, string? subjectId = null, string? subjectRelation = null, string? snapToken = null) { }
+        public string? SnapToken { get; set; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityId { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityType { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string Permission { get; init; }
-        public string? SnapToken { get; init; }
         public string? SubjectId { get; init; }
         public string? SubjectRelation { get; init; }
         public string? SubjectType { get; init; }
@@ -185,11 +185,11 @@ namespace Valtuutus.Core.Engines.Check
             "your compiler.", true)]
         [System.Runtime.CompilerServices.CompilerFeatureRequired("RequiredMembers")]
         public SubjectPermissionRequest() { }
+        public Valtuutus.Core.Data.SnapToken? SnapToken { get; set; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityId { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityType { get; init; }
-        public Valtuutus.Core.Data.SnapToken? SnapToken { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string SubjectId { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
@@ -216,11 +216,11 @@ namespace Valtuutus.Core.Engines.LookupEntity
         public LookupEntityRequest() { }
         [System.Diagnostics.CodeAnalysis.SetsRequiredMembers]
         public LookupEntityRequest(string entityType, string permission, string subjectType, string subjectId) { }
+        public Valtuutus.Core.Data.SnapToken? SnapToken { get; set; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityType { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string Permission { get; init; }
-        public Valtuutus.Core.Data.SnapToken? SnapToken { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string SubjectId { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
@@ -247,13 +247,13 @@ namespace Valtuutus.Core.Engines.LookupSubject
         public LookupSubjectRequest() { }
         [System.Diagnostics.CodeAnalysis.SetsRequiredMembers]
         public LookupSubjectRequest(string entityType, string permission, string subjectType, string entityId) { }
+        public Valtuutus.Core.Data.SnapToken? SnapToken { get; set; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityId { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string EntityType { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string Permission { get; init; }
-        public Valtuutus.Core.Data.SnapToken? SnapToken { get; init; }
         [System.Runtime.CompilerServices.RequiredMember]
         public string SubjectType { get; init; }
     }
