+++ tests/Valtuutus.Data.InMemory.Tests/CheckEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.InMemory.Tests;
 
 [Collection("InMemorySpecs")]
-public sealed class CheckEngineSpecs : DataCheckEngineSpecs
+public sealed class CheckEngineSpecs : BaseCheckEngineSpecs
 {
     public CheckEngineSpecs(InMemoryFixture fixture)
     {
