+++ tests/Valtuutus.Data.InMemory.Tests/DataEngineSpecs.cs
@@ -1,12 +1,12 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Valtuutus.Core;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.InMemory.Tests;
 
 
 [Collection("InMemorySpecs")]
-public sealed class DataEngineSpecs : DataSpecificDataEngineSpecs
+public sealed class DataEngineSpecs : BaseDataEngineSpecs
 {
     protected override IValtuutusDataBuilder AddSpecificProvider(IServiceCollection services)
     {
