+++ tests/Valtuutus.Data.InMemory.Tests/LookupSubjectEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.InMemory.Tests;
 
 [Collection("InMemorySpecs")]
-public sealed class LookupSubjectEngineSpecs : DataLookupSubjectEngineSpecs
+public sealed class LookupSubjectEngineSpecs : BaseLookupSubjectEngineSpecs
 {
 
     public LookupSubjectEngineSpecs(InMemoryFixture fixture)
