+++ tests/Valtuutus.Data.InMemory.Tests/Valtuutus.Data.InMemory.Tests.csproj
@@ -24,7 +24,6 @@
     <ItemGroup>
       <ProjectReference Include="..\..\src\Valtuutus.Core\Valtuutus.Core.csproj" />
       <ProjectReference Include="..\..\src\Valtuutus.Data.InMemory\Valtuutus.Data.InMemory.csproj" />
-      <ProjectReference Include="..\Valtuutus.Data.Tests.Shared\Valtuutus.Data.Tests.Shared.csproj" />
       <ProjectReference Include="..\Valtuutus.Tests.Shared\Valtuutus.Tests.Shared.csproj" />
     </ItemGroup>
