+++ tests/Valtuutus.Data.Postgres.Tests/LookupEntityEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.Postgres.Tests;
 
 [Collection("PostgreSqlSpec")]
-public sealed class LookupEntityEngineSpecs : DataLookupEntityEngineSpecs
+public sealed class LookupEntityEngineSpecs : BaseLookupEntityEngineSpecs
 {
     public LookupEntityEngineSpecs(PostgresFixture fixture)
     {
