+++ tests/Valtuutus.Data.Postgres.Tests/LookupSubjectEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.Postgres.Tests;
 
 [Collection("PostgreSqlSpec")]
-public sealed class LookupSubjectEngineSpecs : DataLookupSubjectEngineSpecs
+public sealed class LookupSubjectEngineSpecs : BaseLookupEntityEngineSpecs
 {
 
     public LookupSubjectEngineSpecs(PostgresFixture fixture)
