+++ tests/Valtuutus.Data.SqlServer.Tests/LookupEntityEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.SqlServer.Tests;
 
 [Collection("SqlServerSpec")]
-public sealed class LookupEntityEngineSpecs : DataLookupEntityEngineSpecs
+public sealed class LookupEntityEngineSpecs : BaseLookupEntityEngineSpecs
 {
 
     public LookupEntityEngineSpecs(SqlServerFixture fixture)
