+++ tests/Valtuutus.Data.SqlServer.Tests/LookupSubjectEngineSpecs.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Valtuutus.Data.Tests.Shared;
+using Valtuutus.Tests.Shared;
 
 namespace Valtuutus.Data.SqlServer.Tests;
 
 [Collection("SqlServerSpec")]
-public sealed class LookupSubjectEngineSpecs : DataLookupSubjectEngineSpecs
+public sealed class LookupSubjectEngineSpecs : BaseLookupSubjectEngineSpecs
 {
 
     public LookupSubjectEngineSpecs(SqlServerFixture fixture)
