--- tests/Valtuutus.Data.Tests.Shared/Valtuutus.Data.Tests.Shared.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFrameworks>$(NetLibVersion);</TargetFrameworks>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\src\Valtuutus.Data.Db\Valtuutus.Data.Db.csproj" />
-      <ProjectReference Include="..\..\src\Valtuutus.Data\Valtuutus.Data.csproj" />
-      <ProjectReference Include="..\Valtuutus.Tests.Shared\Valtuutus.Tests.Shared.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
-      <PackageReference Include="NSubstitute" />
-    </ItemGroup>
-
-</Project>
